requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
Frozen-Flask==1.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pyaml==26.2.1
PyYAML==6.0.3
Werkzeug==3.1.3
